+++ server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/BspServerApi.java
@@ -29,6 +29,9 @@ import ch.epfl.scala.bsp4j.JvmTestEnvironmentParams;
 import ch.epfl.scala.bsp4j.JvmTestEnvironmentResult;
 import ch.epfl.scala.bsp4j.OutputPathsParams;
 import ch.epfl.scala.bsp4j.OutputPathsResult;
+import ch.epfl.scala.bsp4j.PythonBuildServer;
+import ch.epfl.scala.bsp4j.PythonOptionsParams;
+import ch.epfl.scala.bsp4j.PythonOptionsResult;
 import ch.epfl.scala.bsp4j.ResourcesParams;
 import ch.epfl.scala.bsp4j.ResourcesResult;
 import ch.epfl.scala.bsp4j.RunParams;
@@ -59,7 +62,8 @@ public class BspServerApi
         ScalaBuildServer,
         JavaBuildServer,
         CppBuildServer,
-        BazelBuildServer {
+        BazelBuildServer,
+        PythonBuildServer {
 
   private final Supplier<BazelServices> bazelServicesBuilder;
   private BazelBspServerLifetime serverLifetime = null;
@@ -221,6 +225,13 @@ public class BspServerApi
   public CompletableFuture<CppOptionsResult> buildTargetCppOptions(CppOptionsParams params) {
     return runner.handleRequest(
         "buildTargetCppOptions", projectSyncService::buildTargetCppOptions, params);
+  }
+
+  @Override
+  public CompletableFuture<PythonOptionsResult> buildTargetPythonOptions(
+      PythonOptionsParams params) {
+    return runner.handleRequest(
+        "buildTargetPythonOptions", projectSyncService::buildTargetPythonOptions, params);
   }
 
   @Override