+++ server/src/main/java/org/jetbrains/bsp/bazel/server/common/ServerContainer.kt
@@ -18,6 +18,7 @@ import org.jetbrains.bsp.bazel.server.sync.languages.java.JavaLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.java.JdkResolver
 import org.jetbrains.bsp.bazel.server.sync.languages.java.JdkVersionResolver
 import org.jetbrains.bsp.bazel.server.sync.languages.kotlin.KotlinLanguagePlugin
+import org.jetbrains.bsp.bazel.server.sync.languages.python.PythonLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.scala.ScalaLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.thrift.ThriftLanguagePlugin
 import org.jetbrains.bsp.bazel.workspacecontext.WorkspaceContextProvider
@@ -60,8 +61,9 @@ class ServerContainer internal constructor(
             val cppLanguagePlugin = CppLanguagePlugin(bazelPathsResolver)
             val kotlinLanguagePlugin = KotlinLanguagePlugin(javaLanguagePlugin)
             val thriftLanguagePlugin = ThriftLanguagePlugin(bazelPathsResolver)
+            val pythonLanguagePlugin = PythonLanguagePlugin(bazelPathsResolver)
             val languagePluginsService = LanguagePluginsService(
-                scalaLanguagePlugin, javaLanguagePlugin, cppLanguagePlugin, kotlinLanguagePlugin, thriftLanguagePlugin
+                scalaLanguagePlugin, javaLanguagePlugin, cppLanguagePlugin, kotlinLanguagePlugin, thriftLanguagePlugin, pythonLanguagePlugin
             )
             val targetKindResolver = TargetKindResolver()
             val bazelProjectMapper =