+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/languages/LanguagePluginsService.kt
@@ -5,6 +5,8 @@ import org.jetbrains.bsp.bazel.server.sync.languages.cpp.CppLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.cpp.CppModule
 import org.jetbrains.bsp.bazel.server.sync.languages.java.JavaLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.kotlin.KotlinLanguagePlugin
+import org.jetbrains.bsp.bazel.server.sync.languages.python.PythonLanguagePlugin
+import org.jetbrains.bsp.bazel.server.sync.languages.python.PythonModule
 import org.jetbrains.bsp.bazel.server.sync.languages.scala.ScalaLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.thrift.ThriftLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.model.Language
@@ -15,7 +17,8 @@ class LanguagePluginsService(
     val javaLanguagePlugin: JavaLanguagePlugin,
     val cppLanguagePlugin: CppLanguagePlugin,
     val kotlinLanguagePlugin: KotlinLanguagePlugin,
-    private val thriftLanguagePlugin: ThriftLanguagePlugin
+    private val thriftLanguagePlugin: ThriftLanguagePlugin,
+    val pythonLanguagePlugin: PythonLanguagePlugin
 ) {
     private val emptyLanguagePlugin: EmptyLanguagePlugin = EmptyLanguagePlugin()
 
@@ -24,6 +27,7 @@ class LanguagePluginsService(
         javaLanguagePlugin.prepareSync(targetInfos)
         cppLanguagePlugin.prepareSync(targetInfos)
         thriftLanguagePlugin.prepareSync(targetInfos)
+        pythonLanguagePlugin.prepareSync(targetInfos)
     }
 
     fun getPlugin(languages: Set<Language>): LanguagePlugin<*> =
@@ -33,6 +37,7 @@ class LanguagePluginsService(
             languages.contains(Language.KOTLIN) -> kotlinLanguagePlugin
             languages.contains(Language.CPP) -> cppLanguagePlugin
             languages.contains(Language.THRIFT) -> thriftLanguagePlugin
+            languages.contains(Language.PYTHON) -> pythonLanguagePlugin
             else -> emptyLanguagePlugin
         }
 
@@ -40,6 +45,14 @@ class LanguagePluginsService(
         module.languageData?.let {
             when(it) {
                 is CppModule -> it
+                else -> null
+            }
+        }
+
+    fun extractPythonModule(module: Module): PythonModule? =
+        module.languageData?.let {
+            when (it) {
+                is PythonModule -> it
                 else -> null
             }
         }